Move StopWatch into its own file.
