+++ android/src/main/java/com/rnim/rn/audio/AudioRecorderManager.java
@@ -36,48 +36,6 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.IllegalAccessException;
 import java.lang.NoSuchMethodException;
 
-class StopWatch {
-  private long start;
-  private float elapsedTime = 0;
-  private boolean paused = true;
-
-  public StopWatch() {
-  }
-
-  public void start() {
-      start = System.currentTimeMillis();
-      paused = false;
-  }
-
-  public float stop() {
-      if (!paused) {
-          long now = System.currentTimeMillis();
-          elapsedTime += (now - start) / 1000f;
-          paused = true;
-      }
-
-      return elapsedTime;
-  }
-
-  public void reset() {
-      start = 0;
-      elapsedTime = 0;
-      paused = true;
-  }
-
-  public float getTimeSeconds() {
-    float seconds;
-    
-    if (paused) {
-      seconds = elapsedTime;
-    } else {
-      long now = System.currentTimeMillis();
-      seconds = elapsedTime + (now - start) / 1000f;
-    }
-      return seconds;
-  }
-}
-
 class AudioRecorderManager extends ReactContextBaseJavaModule {
 
   private static final String TAG = "ReactNativeAudio";
